fix issues of web.xml
