+++ src/main/webapp/WEB-INF/web.xml
@@ -49,10 +49,10 @@
     <servlet-mapping>
         <servlet-name>fine</servlet-name>
         <url-pattern>/fine</url-pattern>
-
+    </servlet-mapping>
   <servlet>
     <servlet-name>AddComplaint</servlet-name>
-    <servlet-class>com.cops.ntsf.controller.Complaint</servlet-class>
+    <servlet-class>com.cops.ntsf.controller.ComplaintServlet</servlet-class>
   </servlet>
   <servlet-mapping>
     <servlet-name>AddComplaint</servlet-name>
